requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
allennlp==2.10.1
apex==0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.7.9
boto3==1.26.104
botocore==1.29.104
cached-path==1.1.6
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.24.1.1
comm==0.1.2
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
cubinlinker==0.2.2+2.g4de3e99
cuda-python==12.1.0rc5+1.gc7fd38c.dirty
cudf==23.2.0
cugraph==23.2.0
cugraph-dgl==23.2.0
cugraph-service-client==23.2.0
cugraph-service-server==23.2.0
cuml==23.2.0
cupy-cuda12x==12.0.0b3
cycler==0.11.0
cymem==2.0.7
cython==0.29.33
dask==2023.1.1
dask-cuda==23.2.0
dask-cudf==23.2.0
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distributed==2023.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
expecttest==0.1.3
fairscale==0.4.6
fastjsonschema==2.16.3
fastrlock==0.8.1
filelock==3.7.1
flash-attn==0.2.8.dev0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
gitdb==4.0.10
gitpython==3.1.31
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphsurgeon==0.4.6
grpcio==1.51.3
h5py==3.8.0
heapdict==1.0.1
huggingface-hub==0.10.1
hypothesis==5.35.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
jedi==0.18.2
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonnet==0.19.1
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-tensorboard==0.2.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.14.5
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
lit==15.0.7
llvmlite==0.39.1
lmdb==1.4.0
locket==1.0.0
markdown==3.4.1
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.0.1
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.8.1
notebook==6.4.10
numba==0.56.4+1.g9a03de713
numpy==1.22.2
nvidia-dali-cuda110==1.23.0
nvidia-pyindex==1.0.9
nvtx==0.2.5
oauthlib==3.2.2
onnx==1.13.0
opencv==4.6.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
pip==23.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
pluggy==1.0.0
ply==3.11
polygraphy==0.44.2
pooch==1.7.0
preshed==3.0.8
prettytable==3.6.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptxcompiler==0.7.0+27.gbcb4096
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1.dev0+ga6eabc2b.d20230220
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycocotools==2.0+nv0.7.1
pycparser==2.21
pydantic==1.8.2
pygments==2.14.0
pylibcugraph==23.2.0
pylibcugraphops==23.2.0
pylibraft==23.2.0
pynvml==11.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
pytest-rerunfailures==11.1.2
pytest-shard==0.1.2
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-hostlist==1.23.0
pytorch-quantization==2.1.2
pytz==2022.7.1
pyyaml==6.0
pyzmq==25.0.1
raft-dask==23.2.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rich==12.6.0
rmm==23.2.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.2.0
scipy==1.10.1
seaborn==0.12.2
send2trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.18.0
seqeval==1.2.2
setproctitle==1.3.2
setuptools==67.6.1
shortuuid==1.0.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soda-model==0.0.post1.dev1+gb0b71f1.d20230403
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
spacy==3.3.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sphinx-glpi-theme==0.3
srsly==2.4.6
stack-data==0.6.2
strings-udf==23.2.0
sympy==1.11.1
tbb==2021.8.0
tblib==1.7.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6
tensorrt==8.5.3.1
termcolor==1.1.0
terminado==0.17.1
thinc==8.0.17
threadpoolctl==3.1.0
thriftpy2==0.4.16
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu116
torch-tensorrt==1.4.0.dev0
torchaudio==0.12.1+cu116
torchtext==0.13.0a0+fae8e8c
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformer-engine==0.6.0
transformers==4.20.1
treelite==3.1.0
treelite-runtime==3.1.0
triton==2.0.0
typer==0.4.2
types-dataclasses==0.6.6
typing-extensions==4.5.0
ucx-py==0.30.0
uff==0.6.9
urllib3==1.26.14
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
werkzeug==2.2.3
wheel==0.38.4
xdoctest==1.0.2
xgboost==1.7.1
xxhash==3.2.0
yarl==1.8.2
zict==2.2.0
zipp==3.14.0
